backend/requirements.txt
fastapi==0.115.6
sqlmodel==0.0.31
uvicorn[standard]==0.34.0
pyjwt==2.9.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
bcrypt==4.2.0
cohere>=5.20.0
tenacity>=8.2.0

# Phase V - Event-Driven Architecture
aiokafka>=0.10.0
dapr>=1.12.0
APScheduler>=3.10.0
aiosmtplib>=3.0.0
python-dateutil>=2.8.0
